requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
pillow
pydantic[email]
python-multipart
aiofiles
httpx
PyJWT
google-cloud-firestore
firebase-admin
bcrypt
